🚚 Rename CanValidate to CanCurate
